Legger til nye felter i perioder fra infotrygd, og splitter opp tabellene for å vise de kolonner som trengs for gitt stønadstype
