+++ src/frontend/Komponenter/Migrering/InfotrygdPerioder.tsx
@@ -1,0 +1,183 @@
+import { InfotrygdPeriodeMedFlereEndringer } from '../Infotrygd/typer';
+import {
+    formaterNullableIsoDato,
+    formaterNullableMånedÅr,
+    formaterTallMedTusenSkille,
+} from '../../App/utils/formatter';
+import {
+    aktivitetstypeTilTekst,
+    InfotrygdPeriode,
+    Kode,
+    kodeTilTekst,
+    overgangsstønadKodeTilTekst,
+    sakstypeTilTekst,
+} from '../../App/typer/infotrygd';
+import React from 'react';
+import { slåSammenOgSorterPerioder } from '../Infotrygd/grupperInfotrygdperiode';
+import styled from 'styled-components';
+import { Stønadstype } from '../../App/typer/behandlingstema';
+
+const StyledTabell = styled.table``;
+
+const Rad = styled.tr``;
+
+const formatStønadTom = (periode: InfotrygdPeriodeMedFlereEndringer): string => {
+    const stønadTom = formaterNullableMånedÅr(periode.stønadTom) as string;
+    if (periode.kode === Kode.OPPHØRT || periode.kode === Kode.OVERTFØRT_NY_LØSNING) {
+        return `${formaterNullableMånedÅr(periode.opphørsdato)} (${stønadTom})`;
+    } else {
+        return stønadTom;
+    }
+};
+
+const InfotrygdPerioderOvergangsstønad: React.FC<{
+    perioder: InfotrygdPeriodeMedFlereEndringer[];
+}> = ({ perioder }) => {
+    return (
+        <StyledTabell className="tabell">
+            <thead>
+                <tr>
+                    <th>Periode (fom-tom)</th>
+                    <th>Inntektsgrunnlag</th>
+                    <th>Samordningsfradrag</th>
+                    <th>Månedsbeløp</th>
+                    <th>Vedtakstidspunkt</th>
+                    <th>Kode</th>
+                    <th>Sakstype</th>
+                    <th>Aktivitet</th>
+                    <th>Periodetype</th>
+                    <th>Saksbehandler</th>
+                </tr>
+            </thead>
+            <tbody>
+                {perioder.map((periode) => (
+                    <Rad key={`${periode.stønadId}-${periode.vedtakId}`}>
+                        <td>
+                            {formaterNullableMånedÅr(periode.stønadFom)}
+                            {' - '}
+                            {formatStønadTom(periode)}
+                        </td>
+                        <td>{formaterTallMedTusenSkille(periode.inntektsgrunnlag)}</td>
+                        <td>{formaterTallMedTusenSkille(periode.samordningsfradrag)}</td>
+                        <td>{formaterTallMedTusenSkille(periode.månedsbeløp)}</td>
+                        <td>{formaterNullableIsoDato(periode.vedtakstidspunkt)}</td>
+                        <td>
+                            {kodeTilTekst[periode.kode]}{' '}
+                            {periode.initiellKode && `(${kodeTilTekst[periode.initiellKode]})`}
+                        </td>
+                        <td>{sakstypeTilTekst[periode.sakstype]}</td>
+                        <td>
+                            {periode.aktivitetstype &&
+                                aktivitetstypeTilTekst[periode.aktivitetstype]}
+                        </td>
+                        <td>
+                            {periode.kodeOvergangsstønad &&
+                                overgangsstønadKodeTilTekst[periode.kodeOvergangsstønad]}
+                        </td>
+                        <td>{periode.brukerId}</td>
+                    </Rad>
+                ))}
+            </tbody>
+        </StyledTabell>
+    );
+};
+
+const InfotrygdPerioderBarnetilsyn: React.FC<{ perioder: InfotrygdPeriodeMedFlereEndringer[] }> = ({
+    perioder,
+}) => {
+    return (
+        <StyledTabell className="tabell">
+            <thead>
+                <tr>
+                    <th>Periode (fom-tom)</th>
+                    <th>Inntektsgrunnlag</th>
+                    <th>Månedsbeløp</th>
+                    <th>Vedtakstidspunkt</th>
+                    <th>Kode</th>
+                    <th>Sakstype</th>
+                    <th>Saksbehandler</th>
+                </tr>
+            </thead>
+            <tbody>
+                {perioder.map((periode) => (
+                    <Rad key={`${periode.stønadId}-${periode.vedtakId}`}>
+                        <td>
+                            {formaterNullableMånedÅr(periode.stønadFom)}
+                            {' - '}
+                            {formatStønadTom(periode)}
+                        </td>
+                        <td>{formaterTallMedTusenSkille(periode.inntektsgrunnlag)}</td>
+                        <td>{formaterTallMedTusenSkille(periode.månedsbeløp)}</td>
+                        <td>{formaterNullableIsoDato(periode.vedtakstidspunkt)}</td>
+                        <td>
+                            {kodeTilTekst[periode.kode]}{' '}
+                            {periode.initiellKode && `(${kodeTilTekst[periode.initiellKode]})`}
+                        </td>
+                        <td>{sakstypeTilTekst[periode.sakstype]}</td>
+                        <td>{periode.brukerId}</td>
+                    </Rad>
+                ))}
+            </tbody>
+        </StyledTabell>
+    );
+};
+
+const InfotrygdPerioderSkolepenger: React.FC<{ perioder: InfotrygdPeriodeMedFlereEndringer[] }> = ({
+    perioder,
+}) => {
+    return (
+        <StyledTabell className="tabell">
+            <thead>
+                <tr>
+                    <th>Periode (fom-tom)</th>
+                    <th>Månedsbeløp</th>
+                    <th>Engangsbeløp</th>
+                    <th>Vedtakstidspunkt</th>
+                    <th>Kode</th>
+                    <th>Sakstype</th>
+                    <th>Saksbehandler</th>
+                </tr>
+            </thead>
+            <tbody>
+                {perioder.map((periode) => (
+                    <Rad key={`${periode.stønadId}-${periode.vedtakId}`}>
+                        <td>
+                            {formaterNullableMånedÅr(periode.stønadFom)}
+                            {' - '}
+                            {formatStønadTom(periode)}
+                        </td>
+                        <td>{formaterTallMedTusenSkille(periode.månedsbeløp)}</td>
+                        <td>{formaterTallMedTusenSkille(periode.engangsbeløp)}</td>
+                        <td>{formaterNullableIsoDato(periode.vedtakstidspunkt)}</td>
+                        <td>
+                            {kodeTilTekst[periode.kode]}{' '}
+                            {periode.initiellKode && `(${kodeTilTekst[periode.initiellKode]})`}
+                        </td>
+                        <td>{sakstypeTilTekst[periode.sakstype]}</td>
+                        <td>{periode.brukerId}</td>
+                    </Rad>
+                ))}
+            </tbody>
+        </StyledTabell>
+    );
+};
+
+const InfotrygdPerioder: React.FC<{ stønadstype: Stønadstype; perioder: InfotrygdPeriode[] }> = ({
+    stønadstype,
+    perioder,
+}) => {
+    if (perioder.length === 0) {
+        return <>Ingen vedtaksperioder i Infotrygd</>;
+    }
+    const sammenslåttePerioder = slåSammenOgSorterPerioder(perioder);
+    switch (stønadstype) {
+        case Stønadstype.OVERGANGSSTØNAD:
+            return <InfotrygdPerioderOvergangsstønad perioder={sammenslåttePerioder} />;
+        case Stønadstype.BARNETILSYN:
+            return <InfotrygdPerioderBarnetilsyn perioder={sammenslåttePerioder} />;
+        case Stønadstype.SKOLEPENGER:
+            return <InfotrygdPerioderSkolepenger perioder={sammenslåttePerioder} />;
+    }
+};
+
+export default InfotrygdPerioder;
